requirements.txt
antlr4-python3-runtime==4.5.2
beautifulsoup4==4.4.0
certifi==2015.4.28
EasyProcess==0.1.9
Flask==0.10.1
Ghost.py==0.2
gunicorn==19.3.0
htmldom==2.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycurl==7.19.5.1
PyVirtualDisplay==0.1.5
requests==2.7.0
selenium==2.47.1
splinter==0.7.3
urllib3==1.11
virtualenv==13.1.0
Werkzeug==0.10.4
wheel==0.24.0
